+++ guillotina/db/strategies/none.py
@@ -2,12 +2,13 @@ from guillotina import configure
 from guillotina.db.interfaces import IStorage
 from guillotina.db.interfaces import ITransaction
 from guillotina.db.interfaces import ITransactionStrategy
+from guillotina.db.strategies.base import BaseStrategy
 
 
 @configure.adapter(
     for_=(IStorage, ITransaction),
     provides=ITransactionStrategy, name="none")
-class TransactionlessStrategy:
+class TransactionlessStrategy(BaseStrategy):
     """
     Do not handle/detect any conflicts on the database
     """