+++ guillotina/db/transaction_manager.py
@@ -1,104 +1,119 @@
 from asyncio import shield
 from guillotina.db import ROOT_ID
 from guillotina.db.transaction import Transaction
+from guillotina.exceptions import ConflictError
 from guillotina.exceptions import RequestNotFound
 from guillotina.utils import get_authenticated_user_id
 from guillotina.utils import get_current_request
-from queue import LifoQueue
+
+import logging
+
+
+logger = logging.getLogger('guillotina')
 
 
 class TransactionManager(object):
-    """Transaction manager for storing the managed transaction in the
-    current request
-
+    """
+    Transaction manager for storing the managed transaction in the
+    current request object.
     """
 
     def __init__(self, storage):
         # Guillotine Storage
         self._storage = storage
         # Last transaction created
-        self._txn = None
+        self._last_txn = None
         # Pool of transactions
-        self._pool = None
-        self._db_conn = None
-        self.request = None
+        self._last_db_conn = None
 
-    async def root(self):
-        return await self._txn.get(ROOT_ID)
+    async def get_root(self, txn=None):
+        if txn is None:
+            txn = self._last_txn
+        return await txn.get(ROOT_ID)
 
     async def begin(self, request=None):
         """Starts a new transaction.
         """
 
-        self._db_conn = await self._storage.open()
+        db_conn = self._last_db_conn = await self._storage.open()
 
         if request is None:
-            if self.request is None:
-                try:
-                    self.request = get_current_request()
-                except RequestNotFound:
-                    pass
-            request = self.request
+            try:
+                request = get_current_request()
+            except RequestNotFound:
+                pass
 
         user = None
+
+        # if hasattr(request, '_txn'):
+        #     # already has txn registered, close connection on it...
+        #     await self._close_txn(request._txn)
+
+        self._last_txn = txn = Transaction(self, request=request)
+
         if request is not None:
-            request._tm = self  # register it here with request...
+            # register tm and txn with request
+            request._tm = self
+            request._txn = txn
             user = get_authenticated_user_id(request)
-
-        if self._txn is not None:
-            if self._pool is None:
-                self._pool = LifoQueue()
-            # Save the actual transaction and start a new one
-            self._pool.put(self._txn)
-
-        self._txn = txn = Transaction(self, request=request)
 
         # CACHE!!
 
         if user is not None:
             txn.user = user
-        await txn.tpc_begin(self._db_conn)
+        await txn.tpc_begin(db_conn)
 
         return txn
 
-    async def commit(self):
-        return await shield(self._commit())
+    async def commit(self, request=None, txn=None):
+        return await shield(self._commit(request=request, txn=txn))
 
-    async def _commit(self):
+    async def _commit(self, request=None, txn=None):
         """ Commit the last transaction
         """
-        txn = self.get()
+        if txn is None:
+            txn = self.get(request=request)
         if txn is not None:
-            await txn.commit()
-        await self._storage.close(txn._db_conn)
-        self._txn = None
-        self._db_conn = None
-        if self._pool is not None and self._pool.qsize():
-            self._txn = self._pool.get_nowait()
-            self._db_conn = self._txn._db_conn
+            try:
+                await txn.commit()
+            except ConflictError:
+                # we're okay with ConflictError being handled...
+                raise
+            except Exception:
+                logger.error('Error committing transaction {}'.format(txn._tid),
+                             exc_info=True)
+            finally:
+                await self._close_txn(txn)
+        else:
+            await self._close_txn(txn)
 
-    async def abort(self):
-        return await shield(self._abort())
+    async def _close_txn(self, txn):
+        if txn is not None:
+            await self._storage.close(txn._db_conn)
+        if txn == self._last_txn:
+            self._last_txn = None
+            self._last_db_conn = None
 
-    async def _abort(self):
+    async def abort(self, request=None, txn=None):
+        return await shield(self._abort(request=request, txn=txn))
+
+    async def _abort(self, request=None, txn=None):
         """ Abort the last transaction
         """
-        txn = self.get()
+        if txn is None:
+            txn = self.get(request=request)
         if txn is not None:
             await txn.abort()
-        await self._storage.close(txn._db_conn)
-        self._txn = None
-        self._db_conn = None
-        if self._pool is not None and self._pool.qsize():
-            self._txn = self._pool.get_nowait()
-            self._db_conn = self._txn._db_conn
+        await self._close_txn(txn)
 
-    def get(self):
+    def get(self, request=None):
         """Return the current request specific transaction
         """
-        if self._txn:
-            return self._txn
-        if self._pool is not None and self._pool.qsize():
-            self._txn = self._pool.get_nowait()
-            return self._txn
-        return None
+        if request is None:
+            try:
+                request = get_current_request()
+            except RequestNotFound:
+                pass
+        if request is None:
+            return self._last_txn
+        return request._txn