+++ guillotina/tests/test_blob.py
@@ -21,7 +21,7 @@ async def test_create_blob(postgres, guillotina_main):
     blob = Blob(container)
     container.blob = blob
 
-    await request._tm.commit()
+    await request._tm.commit(request)
     await request._tm.begin(request=request)
 
     container = await db.async_get('container')
@@ -29,7 +29,7 @@ async def test_create_blob(postgres, guillotina_main):
     assert blob.resource_zoid == container._p_oid
     await db.async_del('container')
 
-    await request._tm.commit()
+    await request._tm.commit(request)
 
 
 async def test_write_blob_data(postgres, guillotina_main):
@@ -50,7 +50,7 @@ async def test_write_blob_data(postgres, guillotina_main):
     blobfi = blob.open('w')
     await blobfi.async_write(b'foobar')
 
-    await request._tm.commit()
+    await request._tm.commit(request)
     await request._tm.begin(request=request)
 
     container = await db.async_get('container')
@@ -60,7 +60,7 @@ async def test_write_blob_data(postgres, guillotina_main):
 
     await db.async_del('container')
 
-    await request._tm.commit()
+    await request._tm.commit(request)
 
 
 async def test_write_large_blob_data(postgres, guillotina_main):
@@ -83,7 +83,7 @@ async def test_write_large_blob_data(postgres, guillotina_main):
     blobfi = blob.open('w')
     await blobfi.async_write(b'foobar' * multiplier)
 
-    await request._tm.commit()
+    await request._tm.commit(request)
     await request._tm.begin(request=request)
 
     container = await db.async_get('container')
@@ -93,4 +93,4 @@ async def test_write_large_blob_data(postgres, guillotina_main):
 
     await db.async_del('container')
 
-    await request._tm.commit()
+    await request._tm.commit(request)